requirements.txt
streamlit
st-gsheets-connection
pandas
pyzbar
Pillow